Make MapHeader a separate component
