+++ src/client/screens/ProjectScreen.tsx
@@ -2,15 +2,14 @@
 import { useEffect, useState } from "react";
 import { connect } from "react-redux";
 import { Redirect, useParams } from "react-router-dom";
-import Icon from "../components/Icon";
-import { Box, Flex, Label, Select, jsx } from "theme-ui";
-import { IUser, IStaticMetadata } from "../../shared/entities";
-import { setSelectionTool, SelectionTool } from "../actions/districtDrawing";
+import { Flex, jsx } from "theme-ui";
+import { IUser } from "../../shared/entities";
 import { projectDataFetch } from "../actions/projectData";
 import { userFetch } from "../actions/user";
 import "../App.css";
 import CenteredContent from "../components/CenteredContent";
 import Map from "../components/map/Map";
+import MapHeader from "../components/MapHeader";
 import ProjectHeader from "../components/ProjectHeader";
 import ProjectSidebar from "../components/ProjectSidebar";
 import { State } from "../reducers";
@@ -35,60 +34,6 @@ const Main = ({ children }: { readonly children: React.ReactNode }) => {
 
 const MapContainer = ({ children }: { readonly children: React.ReactNode }) => {
   return <Flex sx={{ flexDirection: "column", flex: 1 }}>{children}</Flex>;
-};
-
-const MapHeader = ({
-  label,
-  setMapLabel,
-  metadata,
-  selectionTool
-}: {
-  readonly label?: string;
-  readonly setMapLabel: (label?: string) => void;
-  readonly metadata?: IStaticMetadata;
-  readonly selectionTool: SelectionTool;
-}) => {
-  const options = metadata
-    ? metadata.demographics.map(val => <option key={val.id}>{val.id}</option>)
-    : [];
-  const buttonClassName = (isSelected: boolean) => `map-action ${isSelected ? "selected" : ""}`;
-  return (
-    <Box sx={{ variant: "header.app", backgroundColor: "white" }} className="map-actions">
-      <Box className="actions-left">
-        <Box className="button-group">
-          <button
-            className={buttonClassName(selectionTool === SelectionTool.Default)}
-            onClick={() => store.dispatch(setSelectionTool(SelectionTool.Default))}
-          >
-            <Icon name="hand-pointer" />
-          </button>
-          <button
-            className={buttonClassName(selectionTool === SelectionTool.Rectangle)}
-            onClick={() => store.dispatch(setSelectionTool(SelectionTool.Rectangle))}
-          >
-            <Icon name="draw-square" />
-          </button>
-        </Box>
-      </Box>
-      <Box className="actions-right">
-        <Box className="dropdown">
-          <Label>
-            Label:
-            <Select
-              value={label}
-              onChange={(e: React.ChangeEvent<HTMLSelectElement>) => {
-                const label = e.currentTarget.value;
-                setMapLabel(label);
-              }}
-            >
-              <option></option>
-              {options}
-            </Select>
-          </Label>
-        </Box>
-      </Box>
-    </Box>
-  );
 };
 
 const ProjectScreen = ({ projectData, user, districtDrawing }: StateProps) => {
